pyproject.toml
[tool.poetry]
name = "nonebot-plugin-cogvideox"
version = "0.1.4"
description = "A nonebot plugin for cogvideox"
authors = ["Alpaca <alpaca@alpaca.kim>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-adapter-onebot = "^2.2.1"
zhipuai = "^2.1.5.20250421"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
